Add files via upload
K means clustering algorithms. Its used to group the dataset into k number of predefined non-overlapping clusters
Using the elbow methods, a curve is drawn between “within the sum of squares” (WSS) and the number of clusters.  
The point of the elbow in the curve gives us the optimum number of clusters